+++ ui/icon.c
@@ -90,7 +90,7 @@ UI_GADGET_ICON * ui_add_gadget_icon( UI_WINDOW * wnd, char * text, short x, shor
 
 	icon->width = w;
 	icon->height = h;
-	MALLOC( icon->text, char, strlen( text )+2);
+	MALLOC( icon->text, char, (int)strlen( text )+2);
 	strcpy( icon->text, text );
 	icon->trap_key = k;
 	icon->user_function = f;
